requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-api-commons==1.0.3
django-cors-headers==2.4.0
djangorestframework==3.9.1
gunicorn==19.9.0
pytz==2018.9
typing==3.6.6
